+++ packages/audio-ui/src/index.ts
@@ -1,4 +1,4 @@
-// packages/expo-audio-ui/src/index.ts
+// packages/audio-ui/src/index.ts
 
 export * from './AudioVisualizer/AudioVisualiser.types'
 export * from './AudioVisualizer/AudioVisualizer'